+++ GE2-Test-Unity/Assets/Boid.cs
@@ -13,12 +13,12 @@ public class Boid : MonoBehaviour
 
     //Add fields for length, frequency, start_angle, base_size, multiplier
     public float length;
-    public int freq;
-    public int start_angle;
-    public int base_size;
-    public int multiplier;
+    public float freq;
+    public float start_angle;
+    public float base_size;
+    public float multiplier;
 
-    //move to diif dest
+    //move to dif dest
     public GameObject targetCube;
 
     [Range(0.0f, 10.0f)]
